ZPM and other stuff

+++ src/main/java/mrjake/aunis/gui/container/zpmhub/ZPMHubContainerGuiUpdate.java
@@ -8,9 +8,9 @@ public class ZPMHubContainerGuiUpdate extends State {
 	public ZPMHubContainerGuiUpdate() {}
 
 	public int zpmsCount;
-	public int energyTransferedLastTick;
+	public long energyTransferedLastTick;
 
-	public ZPMHubContainerGuiUpdate(int zpmsCount, int energyTransferedLastTick) {
+	public ZPMHubContainerGuiUpdate(int zpmsCount, long energyTransferedLastTick) {
 		this.zpmsCount = zpmsCount;
 		this.energyTransferedLastTick = energyTransferedLastTick;
 	}
@@ -18,13 +18,13 @@ public class ZPMHubContainerGuiUpdate extends State {
 	@Override
 	public void toBytes(ByteBuf buf) {
 		buf.writeInt(zpmsCount);
-		buf.writeInt(energyTransferedLastTick);
+		buf.writeLong(energyTransferedLastTick);
 	}
 
 	@Override
 	public void fromBytes(ByteBuf buf) {
 		zpmsCount = buf.readInt();
-		energyTransferedLastTick = buf.readInt();
+		energyTransferedLastTick = buf.readLong();
 	}
 
 }
