+++ src/main/java/mrjake/aunis/renderer/energy/ZPMHubRendererState.java
@@ -9,7 +9,7 @@ public class ZPMHubRendererState extends State {
 
     public ZPMHubRendererState() {}
 
-    public ZPMHubRendererState(BlockPos pos, long tick, boolean isPutting, int zpmsCount, int zpmAnimated, int energyTransferedLastTick) {
+    public ZPMHubRendererState(BlockPos pos, long tick, boolean isPutting, int zpmsCount, int zpmAnimated, long energyTransferedLastTick) {
         this.pos = pos;
         this.animationStart = tick;
         this.isPutting = isPutting;
@@ -18,7 +18,7 @@ public class ZPMHubRendererState extends State {
         this.energyTransferedLastTick = energyTransferedLastTick;
     }
 
-    public ZPMHubRendererState initClient(BlockPos pos, long tick, boolean isPutting, int zpmsCount, int zpmAnimated, int energyTransferedLastTick) {
+    public ZPMHubRendererState initClient(BlockPos pos, long tick, boolean isPutting, int zpmsCount, int zpmAnimated, long energyTransferedLastTick) {
         this.pos = pos;
         this.animationStart = tick;
         this.isPutting = isPutting;
@@ -42,7 +42,7 @@ public class ZPMHubRendererState extends State {
     public boolean isPutting;
     public int zpmsCount;
     public int zpmAnimated;
-    public int energyTransferedLastTick;
+    public long energyTransferedLastTick;
 
     public BiomeOverlayEnum getBiomeOverlay() {
         return BiomeOverlayEnum.NORMAL;
@@ -54,7 +54,7 @@ public class ZPMHubRendererState extends State {
         buf.writeBoolean(isPutting);
         buf.writeInt(zpmsCount);
         buf.writeInt(zpmAnimated);
-        buf.writeInt(energyTransferedLastTick);
+        buf.writeLong(energyTransferedLastTick);
     }
 
     public void fromBytes(ByteBuf buf) {
@@ -63,6 +63,6 @@ public class ZPMHubRendererState extends State {
         isPutting = buf.readBoolean();
         zpmsCount = buf.readInt();
         zpmAnimated = buf.readInt();
-        energyTransferedLastTick = buf.readInt();
+        energyTransferedLastTick = buf.readLong();
     }
 }
